+++ packages/terra-table/src/Table.jsx
@@ -3,34 +3,41 @@ import React, {
 } from 'react';
 import PropTypes from 'prop-types';
 import { injectIntl } from 'react-intl';
+import { v4 as uuidv4 } from 'uuid';
 import classNames from 'classnames/bind';
 import ResizeObserver from 'resize-observer-polyfill';
 import ThemeContext from 'terra-theme-context';
 import VisuallyHiddenText from 'terra-visually-hidden-text';
+import Section from './subcomponents/Section';
 import ColumnHeader from './subcomponents/ColumnHeader';
 import ColumnContext from './utils/ColumnContext';
 import { columnShape } from './proptypes/columnShape';
 import ERRORS from './utils/constants';
 import GridContext, { GridConstants } from './utils/GridContext';
-import Row from './subcomponents/Row';
 import rowShape from './proptypes/rowShape';
 import validateRowHeaderIndex from './proptypes/validators';
 
 import styles from './Table.module.scss';
+import sectionShape from './proptypes/sectionShape';
 
 const cx = classNames.bind(styles);
 
 const propTypes = {
   /**
-   * String that will be used to identify the table. If multiple tables are on the same page, each table should have
-   * a unique id.
+   * Unique id used to identify the table.
    */
   id: PropTypes.string.isRequired,
 
   /**
-  * Data for content in the body of the table. Rows will be rendered in the order given.
+  * Data for content in the body of the table when sections do not exist. Rows will be rendered in the order given.
+  * The sections prop will text precedence over this prop.
   */
   rows: PropTypes.arrayOf(rowShape),
+
+  /**
+  * Data for content in the body of the table. Sections will be rendered in the order given.
+  */
+  sections: PropTypes.arrayOf(sectionShape),
 
   /**
    * String that identifies the element (or elements) that labels the table.
@@ -118,6 +125,11 @@ const propTypes = {
   onColumnSelect: PropTypes.func,
 
   /**
+   * Function that is called when a collapsible section is selected. Parameters: `onSectionSelect(sectionId)`
+   */
+  onSectionSelect: PropTypes.func,
+
+  /*
    * Callback function that is called when the row selection column header is selected. Parameters:
    *  @param {string} columnId columnId
    */
@@ -170,6 +182,7 @@ function Table(props) {
     isActiveColumnResizing,
     columnResizeIncrement,
     rows,
+    sections,
     pinnedColumns,
     overflowColumns,
     onColumnResize,
@@ -178,6 +191,7 @@ function Table(props) {
     rowHeight,
     onColumnSelect,
     onCellSelect,
+    onSectionSelect,
     onRowSelect,
     onRowSelectionHeaderSelect,
     hasSelectableRows,
@@ -238,6 +252,32 @@ function Table(props) {
 
   const displayedColumns = (hasSelectableRows ? [tableRowSelectionColumn] : []).concat(pinnedColumns).concat(overflowColumns);
   const [tableColumns, setTableColumns] = useState(displayedColumns.map((column) => initializeColumn(column)));
+
+  const defaultSectionRef = useRef(uuidv4());
+
+  // Create section array from props
+  const tableSections = useMemo(() => {
+    if (sections) {
+      return [...sections];
+    }
+
+    return [{ id: defaultSectionRef.current, rows }];
+  }, [rows, sections]);
+
+  // Calculate total table row count
+  const tableSectionReducer = (rowCount, currentSection) => {
+    if (currentSection.id !== defaultSectionRef.current) {
+      // eslint-disable-next-line no-param-reassign
+      currentSection.sectionRowIndex = rowCount + 1;
+      return rowCount + currentSection.rows.length + 1;
+    }
+
+    // eslint-disable-next-line no-param-reassign
+    currentSection.sectionRowIndex = 0;
+    return rowCount + currentSection.rows.length;
+  };
+  const tableRowCount = tableSections.reduce(tableSectionReducer, 1);
+
   // -------------------------------------
   // functions
 
@@ -271,11 +311,12 @@ function Table(props) {
   // useEffect for row updates
   useEffect(() => {
     const previousSelectedRows = [...selectedRows.current];
-    selectedRows.current = rows.filter((row) => row.isSelected).map(row => (row.id));
+    const selectableRows = tableSections.flatMap(section => (section.rows.map(row => (row))));
+    selectedRows.current = selectableRows.filter((row) => row.isSelected).map(row => (row.id));
 
     if (previousSelectedRows.length > 0 && selectedRows.current.length === 0) {
       setRowSelectionAriaLiveMessage(intl.formatMessage({ id: 'Terra.table.all-rows-unselected' }));
-    } else if (selectedRows.current.length === rows.length) {
+    } else if (selectedRows.current.length === selectableRows.length) {
       setRowSelectionAriaLiveMessage(intl.formatMessage({ id: 'Terra.table.all-rows-selected' }));
     } else {
       const rowSelectionsAdded = selectedRows.current.filter(row => !previousSelectedRows.includes(row));
@@ -283,16 +324,14 @@ function Table(props) {
       let selectionUpdateAriaMessage = '';
 
       if (rowSelectionsAdded.length === 1) {
-        const newRowIndex = rows.findIndex(row => row.id === rowSelectionsAdded[0]);
-        const selectedRowLabel = rows[newRowIndex].ariaLabel || newRowIndex + 2; // Accounts for header row and zero-based index
+        const selectedRowLabel = tableRef.current.querySelector(`tr[data-row-id='${rowSelectionsAdded[0]}']`).getAttribute('aria-rowindex');
         selectionUpdateAriaMessage = intl.formatMessage({ id: 'Terra.table.row-selection-template' }, { row: selectedRowLabel });
       } else if (rowSelectionsAdded.length > 1) {
         selectionUpdateAriaMessage = intl.formatMessage({ id: 'Terra.table.multiple-rows-selected' }, { rowCount: rowSelectionsAdded.length });
       }
 
       if (rowSelectionsRemoved.length === 1) {
-        const removedRowIndex = rows.findIndex(row => row.id === rowSelectionsRemoved[0]);
-        const unselectedRowLabel = rows[removedRowIndex].ariaLabel || removedRowIndex + 2; // Accounts for header row and zero-based index
+        const unselectedRowLabel = tableRef.current.querySelector(`tr[data-row-id='${rowSelectionsRemoved[0]}']`).getAttribute('data-row-id');
         selectionUpdateAriaMessage += intl.formatMessage({ id: 'Terra.table.row-selection-cleared-template' }, { row: unselectedRowLabel });
       } else if (rowSelectionsRemoved.length > 1) {
         selectionUpdateAriaMessage += intl.formatMessage({ id: 'Terra.table.multiple-rows-unselected' }, { rowCount: rowSelectionsRemoved.length });
@@ -302,7 +341,7 @@ function Table(props) {
         setRowSelectionAriaLiveMessage(selectionUpdateAriaMessage);
       }
     }
-  }, [intl, rows]);
+  }, [intl, tableSections]);
 
   // useEffect for row displayed columns
   useEffect(() => {
@@ -454,6 +493,7 @@ function Table(props) {
         role={gridContext.role}
         aria-labelledby={ariaLabelledBy}
         aria-label={ariaLabel}
+        aria-rowcount={tableRowCount}
         className={cx('table', theme.className, { headerless: !hasColumnHeaders })}
         {...(activeIndex != null && { onMouseUp, onMouseMove, onMouseLeave: onMouseUp })}
       >
@@ -468,6 +508,7 @@ function Table(props) {
           </colgroup>
 
           <ColumnHeader
+            tableId={id}
             isActiveColumnResizing={isActiveColumnResizing}
             activeColumnIndex={activeColumnIndex}
             columns={tableColumns}
@@ -479,24 +520,26 @@ function Table(props) {
             onColumnSelect={handleColumnSelect}
             onResizeHandleChange={onResizeHandleChange}
           />
-          <tbody>
-            {rows.map((row, index) => (
-              <Row
-                rowIndex={index + 1}
-                key={row.id}
-                height={rowHeight}
-                id={row.id}
-                cells={row.cells}
-                ariaLabel={row.ariaLabel}
-                hasRowSelection={hasSelectableRows}
-                displayedColumns={displayedColumns}
-                rowHeaderIndex={rowHeaderIndex}
-                onCellSelect={isGridContext || hasSelectableRows ? handleCellSelection : undefined}
-                isSelected={row.isSelected}
-                isTableStriped={isStriped}
-              />
-            ))}
-          </tbody>
+          {tableSections.map((section) => (
+            <Section
+              id={section.id}
+              tableId={id}
+              key={section.id}
+              sectionRowIndex={section.sectionRowIndex}
+              isCollapsible={section.isCollapsible}
+              isCollapsed={section.isCollapsed}
+              isHidden={section.id === defaultSectionRef.current}
+              isTableStriped={isStriped}
+              text={section.text}
+              rows={section.rows}
+              rowHeight={rowHeight}
+              hasRowSelection={hasSelectableRows}
+              displayedColumns={displayedColumns}
+              rowHeaderIndex={rowHeaderIndex}
+              onCellSelect={isGridContext || hasSelectableRows ? handleCellSelection : undefined}
+              onSectionSelect={onSectionSelect}
+            />
+          ))}
         </ColumnContext.Provider>
       </table>
       <VisuallyHiddenText aria-live="polite" text={rowSelectionModeAriaLiveMessage} />